Trait ct_sdk::ct::sdk::CTSignedRequest

pub trait CTSignedRequest<'a> {
    fn presigned(
        &mut self, 
        creds: &AwsCredentials, 
        date: &Option<String>
    ) -> (String, String);

    fn add_header_raw(&mut self, key: &str, value: &str);

    fn gen_url(&mut self) -> String;

    fn url(&mut self) -> String;
}

A trait to abstract the idea of generate a pre-signed Url for an S3 object from a SignedRequest.

Required Methods

fn presigned(
&mut self,
creds: &AwsCredentials,
date: &Option<String>
) -> (String, String)

Pre-signed SignedRequest use SignV2

fn add_header_raw(&mut self, key: &str, value: &str)

Add a value to the array of headers for the specified key. Headers are kept sorted by key name for use at signing (BTreeMap). But in the query the content don`t need to lowercase.(RFC2616)

fn gen_url(&mut self) -> String

Generate Url from a SignedRequest

fn url(&mut self) -> String